requirements.txt
# Core dependencies for Resume Parser
streamlit==1.29.0
PyMuPDF==1.23.8
python-docx==1.1.0
spacy==3.7.2
pandas==2.1.4
numpy==1.26.2
openpyxl==3.1.2

# For regex and text processing
regex==2023.12.25

# For NLP model
# Run: python -m spacy download en_core_web_sm
